+++ apps/sim/lib/table/service.ts
@@ -8,7 +8,7 @@
  */
 
 import { db } from '@sim/db'
-import { tableJobs, tableRowExecutions, userTableDefinitions, userTableRows } from '@sim/db/schema'
+import { tableJobs, userTableDefinitions, userTableRows } from '@sim/db/schema'
 import { createLogger } from '@sim/logger'
 import { getPostgresErrorCode, toError } from '@sim/utils/errors'
 import { generateId } from '@sim/utils/id'
@@ -19,10 +19,8 @@ import {
   desc,
   eq,
   gt,
-  gte,
   inArray,
   isNull,
-  lt,
   lte,
   ne,
   notInArray,
@@ -41,12 +39,31 @@ import {
   withGeneratedColumnIds,
 } from './column-keys'
 import { COLUMN_TYPES, NAME_PATTERN, TABLE_LIMITS, USER_TABLE_ROWS_SQL_NAME } from './constants'
-import { areGroupDepsSatisfied } from './deps'
 import { CSV_MAX_BATCH_SIZE } from './import'
-import { keyBetween, nKeysBetween } from './order-key'
+import { nKeysBetween } from './order-key'
 import { type DbExecutor, type DbTransaction, withSeqscanOff } from './planner'
+import {
+  applyExecutionsPatch,
+  deriveExecClearsForDataPatch,
+  loadExecutionsByRow,
+  loadExecutionsForRow,
+  stripGroupExecutions,
+  writeExecutionsPatch,
+} from './rows/executions'
+import {
+  acquireRowOrderLock,
+  deleteOrderedRow,
+  deleteOrderedRowsByIds,
+  insertOrderedRow,
+  nextRowPosition,
+  reserveBatchPositions,
+  reserveInsertPosition,
+  resolveBatchInsertOrderKeys,
+  resolveInsertOrderKey,
+} from './rows/ordering'
 import { buildFilterClause, buildSortClause, escapeLikePattern } from './sql'
 import { fireTableTrigger } from './trigger'
+import { scaledStatementTimeoutMs, setTableTxTimeouts } from './tx'
 import type {
   AddWorkflowGroupData,
   BatchInsertData,
@@ -116,27 +133,6 @@ export class TableConflictError extends Error {
 export type TableScope = 'active' | 'archived' | 'all'
 
 /**
- * Sets per-transaction Postgres timeouts via `SET LOCAL`.
- *
- * `lock_timeout` is the critical one: without it, a waiter inherits the full
- * `statement_timeout` clock, so one stuck writer can drain the pool.
- *
- * Safe under pgBouncer transaction pooling — `SET LOCAL` is transaction-scoped
- * and cleared at COMMIT/ROLLBACK before the session returns to the pool.
- */
-async function setTableTxTimeouts(
-  trx: DbTransaction,
-  opts?: { statementMs?: number; lockMs?: number; idleMs?: number }
-) {
-  const s = opts?.statementMs ?? 10_000
-  const l = opts?.lockMs ?? 3_000
-  const i = opts?.idleMs ?? 5_000
-  await trx.execute(sql.raw(`SET LOCAL statement_timeout = '${s}ms'`))
-  await trx.execute(sql.raw(`SET LOCAL lock_timeout = '${l}ms'`))
-  await trx.execute(sql.raw(`SET LOCAL idle_in_transaction_session_timeout = '${i}ms'`))
-}
-
-/**
  * Serializes schema/metadata read-modify-writes for a single table so
  * concurrent mutators can't clobber each other's `schema` JSONB
  * (last-writer-wins). Takes a transaction-scoped advisory lock keyed on
@@ -167,50 +163,6 @@ async function withLockedTable<T>(
     }
     return mutate(table, trx)
   })
-}
-
-/**
- * Starting `position` for an append import — `max(position) + 1`, or 0 when empty. Read once,
- * unlocked, before streaming: the import worker is the table's sole writer, so it can assign
- * contiguous positions from this offset without per-batch position scans.
- */
-export async function nextImportStartPosition(tableId: string): Promise<number> {
-  const [{ maxPos }] = await db
-    .select({
-      maxPos: sql<number>`coalesce(max(${userTableRows.position}), -1)`.mapWith(Number),
-    })
-    .from(userTableRows)
-    .where(eq(userTableRows.tableId, tableId))
-  return maxPos + 1
-}
-
-/**
- * Append anchor `order_key` for an import — `max(order_key)`, or null when empty. Read once,
- * unlocked, before streaming (the import worker is the table's sole writer); each batch threads
- * the previous batch's last key forward so no per-batch max scan is needed.
- */
-export async function nextImportStartOrderKey(tableId: string): Promise<string | null> {
-  return maxOrderKey(db, tableId)
-}
-
-const TIMEOUT_CAP_MS = 10 * 60_000
-
-/**
- * Scales `statement_timeout` to the expected row-count work.
- *
- * Bulk operations that rewrite JSONB or cascade row triggers (e.g.
- * `replaceTableRows`, `deleteColumn`, `renameColumn`) scale roughly linearly
- * with row count. A fixed cap would regress large-table users who never saw a
- * timeout before `SET LOCAL` was introduced. This helper picks
- * `max(baseMs, rowCount * perRowMs)`, capped at 10 minutes so a single
- * runaway transaction cannot indefinitely pin a pool connection.
- */
-function scaledStatementTimeoutMs(
-  rowCount: number,
-  opts: { baseMs: number; perRowMs: number }
-): number {
-  const safeRowCount = Math.max(0, rowCount)
-  return Math.min(TIMEOUT_CAP_MS, Math.max(opts.baseMs, safeRowCount * opts.perRowMs))
 }
 
 /**
@@ -1084,604 +1036,6 @@ export async function restoreTable(tableId: string, requestId: string): Promise<
   }
 
   logger.info(`[${requestId}] Restored table ${tableId} as "${attemptedRestoreName}"`)
-}
-
-/**
- * Loads `tableRowExecutions` rows for the given row ids and groups them into a
- * `Map<rowId, RowExecutions>` suitable for plugging into `TableRow.executions`.
- */
-async function loadExecutionsByRow(
-  trx: DbOrTx,
-  rowIds: Iterable<string>
-): Promise<Map<string, RowExecutions>> {
-  const ids = Array.from(new Set(rowIds))
-  const result = new Map<string, RowExecutions>()
-  if (ids.length === 0) return result
-  const rows = await trx
-    .select()
-    .from(tableRowExecutions)
-    .where(inArray(tableRowExecutions.rowId, ids))
-  for (const r of rows) {
-    const existing = result.get(r.rowId) ?? {}
-    const meta: RowExecutionMetadata = {
-      status: r.status as RowExecutionMetadata['status'],
-      executionId: r.executionId ?? null,
-      jobId: r.jobId ?? null,
-      workflowId: r.workflowId,
-      error: r.error ?? null,
-      ...(r.runningBlockIds && r.runningBlockIds.length > 0
-        ? { runningBlockIds: r.runningBlockIds }
-        : {}),
-      ...(r.blockErrors && Object.keys(r.blockErrors as Record<string, string>).length > 0
-        ? { blockErrors: r.blockErrors as Record<string, string> }
-        : {}),
-      ...(r.cancelledAt ? { cancelledAt: r.cancelledAt.toISOString() } : {}),
-    }
-    existing[r.groupId] = meta
-    result.set(r.rowId, existing)
-  }
-  return result
-}
-
-/** Convenience: load executions for one row, returning `{}` when missing. */
-async function loadExecutionsForRow(trx: DbOrTx, rowId: string): Promise<RowExecutions> {
-  const byRow = await loadExecutionsByRow(trx, [rowId])
-  return byRow.get(rowId) ?? {}
-}
-
-/**
- * Serializes writers that assign `position` for the same table. The row-count
- * trigger (migration 0198) serializes capacity via a row lock on
- * `user_table_definitions`, but it fires AFTER INSERT, so two concurrent
- * auto-positioned inserts could read the same snapshot and assign the same
- * position (the `(table_id, position)` index is non-unique). This advisory lock
- * restores per-table serialization. Released at COMMIT/ROLLBACK.
- */
-async function acquireRowOrderLock(trx: DbTransaction, tableId: string) {
-  await trx.execute(
-    sql`SELECT pg_advisory_xact_lock(hashtextextended(${`user_table_rows_pos:${tableId}`}, 0))`
-  )
-}
-
-/** Next append position for a table (max(position) + 1, or 0 if empty). */
-async function nextRowPosition(trx: DbTransaction, tableId: string): Promise<number> {
-  const [{ maxPos }] = await trx
-    .select({
-      maxPos: sql<number>`coalesce(max(${userTableRows.position}), -1)`.mapWith(Number),
-    })
-    .from(userTableRows)
-    .where(eq(userTableRows.tableId, tableId))
-  return maxPos + 1
-}
-
-/** Largest `order_key` for a table, or `null` when empty — the append anchor for new keys. */
-async function maxOrderKey(executor: DbOrTx, tableId: string): Promise<string | null> {
-  const [{ maxKey }] = await executor
-    .select({ maxKey: sql<string | null>`max(${userTableRows.orderKey})` })
-    .from(userTableRows)
-    .where(eq(userTableRows.tableId, tableId))
-  return maxKey ?? null
-}
-
-/** Shifts every row at or after `position` up by one (`position + 1`). */
-async function shiftRowsUpFrom(trx: DbTransaction, tableId: string, position: number) {
-  await trx
-    .update(userTableRows)
-    .set({ position: sql`position + 1` })
-    .where(and(eq(userTableRows.tableId, tableId), gte(userTableRows.position, position)))
-}
-
-/** Shifts every row after `position` down by one (`position - 1`). */
-async function shiftRowsDownAfter(trx: DbTransaction, tableId: string, position: number) {
-  await trx
-    .update(userTableRows)
-    .set({ position: sql`position - 1` })
-    .where(and(eq(userTableRows.tableId, tableId), gt(userTableRows.position, position)))
-}
-
-/**
- * Reserves the `position` for a single inserted row and returns where to INSERT.
- * Acquires the row-order lock, then opens a slot at `requestedPosition` (shifting
- * the occupant + tail up) or computes the append position. Caller runs inside a
- * transaction.
- */
-async function reserveInsertPosition(
-  trx: DbTransaction,
-  tableId: string,
-  requestedPosition?: number
-): Promise<number> {
-  await acquireRowOrderLock(trx, tableId)
-  if (requestedPosition === undefined) {
-    return nextRowPosition(trx, tableId)
-  }
-  const [existing] = await trx
-    .select({ id: userTableRows.id })
-    .from(userTableRows)
-    .where(and(eq(userTableRows.tableId, tableId), eq(userTableRows.position, requestedPosition)))
-    .limit(1)
-  if (existing) {
-    await shiftRowsUpFrom(trx, tableId, requestedPosition)
-  }
-  return requestedPosition
-}
-
-/**
- * Reserves positions for a batch of `count` rows. Opens each requested slot
- * (ascending, preserving prior gaps) and returns the requested positions in
- * original order; otherwise returns a contiguous append range.
- */
-async function reserveBatchPositions(
-  trx: DbTransaction,
-  tableId: string,
-  count: number,
-  requestedPositions?: number[]
-): Promise<number[]> {
-  await acquireRowOrderLock(trx, tableId)
-  if (requestedPositions && requestedPositions.length > 0) {
-    for (const pos of [...requestedPositions].sort((a, b) => a - b)) {
-      await shiftRowsUpFrom(trx, tableId, pos)
-    }
-    return requestedPositions
-  }
-  const start = await nextRowPosition(trx, tableId)
-  return Array.from({ length: count }, (_, i) => start + i)
-}
-
-/**
- * Recompacts row positions to be contiguous after a bulk delete. With
- * `minDeletedPos`, only rows at/after it are re-numbered; single-row deletes use
- * the cheaper {@link shiftRowsDownAfter}.
- */
-async function compactPositions(trx: DbTransaction, tableId: string, minDeletedPos?: number) {
-  if (minDeletedPos === undefined) {
-    await trx.execute(sql`
-      UPDATE user_table_rows t
-      SET position = r.new_pos
-      FROM (
-        SELECT id, ROW_NUMBER() OVER (ORDER BY position) - 1 AS new_pos
-        FROM user_table_rows
-        WHERE table_id = ${tableId}
-      ) r
-      WHERE t.id = r.id AND t.table_id = ${tableId} AND t.position != r.new_pos
-    `)
-    return
-  }
-  await trx.execute(sql`
-    UPDATE user_table_rows t
-    SET position = r.new_pos
-    FROM (
-      SELECT id, ${minDeletedPos}::int + ROW_NUMBER() OVER (ORDER BY position) - 1 AS new_pos
-      FROM user_table_rows
-      WHERE table_id = ${tableId} AND position >= ${minDeletedPos}
-    ) r
-    WHERE t.id = r.id AND t.table_id = ${tableId} AND t.position != r.new_pos
-  `)
-}
-
-/** A row value ready to INSERT into `user_table_rows`, with its assigned order. */
-export interface OrderedRowValue {
-  id: string
-  tableId: string
-  workspaceId: string
-  data: RowData
-  position: number
-  orderKey: string
-  createdAt: Date
-  updatedAt: Date
-  createdBy?: string
-}
-
-/**
- * Builds INSERT values for a contiguous run of rows, assigning sequential
- * positions `startPosition + i` and the supplied `orderKeys[i]`. Centralizes
- * row assignment for callers that write a fresh ordered run (e.g. the copilot
- * tool's replace-all write). `orderKeys` must be index-aligned with `rows` —
- * mint them once for the whole run with {@link nKeysBetween}.
- */
-export function buildOrderedRowValues(opts: {
-  tableId: string
-  workspaceId: string
-  rows: RowData[]
-  startPosition: number
-  orderKeys: string[]
-  now: Date
-  createdBy?: string
-  makeId: () => string
-}): OrderedRowValue[] {
-  const { tableId, workspaceId, rows, startPosition, orderKeys, now, createdBy, makeId } = opts
-  return rows.map((data, i) => ({
-    id: makeId(),
-    tableId,
-    workspaceId,
-    data,
-    position: startPosition + i,
-    orderKey: orderKeys[i],
-    createdAt: now,
-    updatedAt: now,
-    ...(createdBy ? { createdBy } : {}),
-  }))
-}
-
-/**
- * Computes the fractional `order_key` for a row inserted at the integer
- * `requestedPosition` (or appended when omitted). Used by position-based callers
- * (mothership tool, v1 API, undo position-fallback, transient old clients).
- *
- * The neighbor at slot `s` is resolved differently per flag state:
- * - **off**: `WHERE position = s` (positions are contiguous, so the row at
- *   position `s` is the `s`-th row — an indexed O(1) lookup).
- * - **on**: the `s`-th row in `order_key, id` order (`OFFSET s`) — positions are
- *   gappy and non-authoritative, so `position = s` would miss; the visual
- *   ordinal is the key's ordinal. O(s), acceptable for these low-volume callers.
- *
- * Caller holds the row-order lock.
- */
-async function resolveInsertOrderKey(
-  trx: DbTransaction,
-  tableId: string,
-  requestedPosition?: number
-): Promise<string> {
-  const orderKeyAtSlot = async (slot: number): Promise<string | null> => {
-    if (slot < 0) return null
-    if (isTablesFractionalOrderingEnabled) {
-      const [r] = await trx
-        .select({ orderKey: userTableRows.orderKey })
-        .from(userTableRows)
-        .where(eq(userTableRows.tableId, tableId))
-        .orderBy(asc(userTableRows.orderKey), asc(userTableRows.id))
-        .limit(1)
-        .offset(slot)
-      return r?.orderKey ?? null
-    }
-    const [r] = await trx
-      .select({ orderKey: userTableRows.orderKey })
-      .from(userTableRows)
-      .where(and(eq(userTableRows.tableId, tableId), eq(userTableRows.position, slot)))
-      .limit(1)
-    return r?.orderKey ?? null
-  }
-  if (requestedPosition === undefined) {
-    return keyBetween(await maxOrderKey(trx, tableId), null)
-  }
-  const lo = await orderKeyAtSlot(requestedPosition - 1)
-  const hi = await orderKeyAtSlot(requestedPosition)
-  return keyBetween(lo, hi)
-}
-
-/**
- * Resolves the `order_key` for an insert expressed by an anchor row id —
- * `afterRowId` (place directly after) or `beforeRowId` (directly before). Finds
- * the anchor and its adjacent key via the `(table_id, order_key, id)` index
- * (O(1)) and mints a key between them. Also returns a legacy integer `position`
- * (anchor's position ±) so the flag-off shift path still works. Caller holds the
- * row-order lock.
- */
-async function resolveInsertByNeighbor(
-  trx: DbTransaction,
-  tableId: string,
-  afterRowId?: string,
-  beforeRowId?: string
-): Promise<{ orderKey: string; position: number }> {
-  const anchorId = afterRowId ?? beforeRowId!
-  const [anchor] = await trx
-    .select({ orderKey: userTableRows.orderKey, position: userTableRows.position })
-    .from(userTableRows)
-    .where(and(eq(userTableRows.tableId, tableId), eq(userTableRows.id, anchorId)))
-    .limit(1)
-  // The client targets a specific neighbor; a missing one (concurrent delete /
-  // stale view) is an error, not a silent insert at the front.
-  if (!anchor) throw new Error(`Row not found: ${anchorId}`)
-  const anchorKey = anchor.orderKey ?? null
-  // A null key on the anchor means the table isn't backfilled. With the flag on
-  // (key is authoritative) the adjacent-key lookup below can't work — fail
-  // loudly rather than mint a wrong key. Flag off keeps `position` authoritative,
-  // so a best-effort key here is fine (the backfill re-keys before the flip).
-  if (anchorKey === null && isTablesFractionalOrderingEnabled) {
-    throw new Error(`Row ${anchorId} has no order_key yet (table not backfilled)`)
-  }
-
-  if (afterRowId) {
-    // hi = the smallest key strictly GREATER than the anchor key. Comparing keys
-    // (not the `(order_key, id)` row tuple) skips past any sibling that shares the
-    // anchor's key, so `keyBetween` always gets strictly-ordered bounds and can't
-    // throw on a stray duplicate. Identical to the row tuple when keys are distinct.
-    // A null anchorKey (flag off, un-backfilled) has no key to compare — leave the
-    // upper bound open, matching the prior best-effort behavior.
-    let nextKey: string | null = null
-    if (anchorKey !== null) {
-      const [next] = await trx
-        .select({ orderKey: userTableRows.orderKey })
-        .from(userTableRows)
-        .where(and(eq(userTableRows.tableId, tableId), gt(userTableRows.orderKey, anchorKey)))
-        .orderBy(asc(userTableRows.orderKey))
-        .limit(1)
-      nextKey = next?.orderKey ?? null
-    }
-    return {
-      orderKey: keyBetween(anchorKey, nextKey),
-      position: anchor.position + 1,
-    }
-  }
-
-  // beforeRowId: lo = the largest key strictly LESS than the anchor key (distinct,
-  // same rationale as the afterRowId branch above).
-  let prevKey: string | null = null
-  if (anchorKey !== null) {
-    const [prev] = await trx
-      .select({ orderKey: userTableRows.orderKey })
-      .from(userTableRows)
-      .where(and(eq(userTableRows.tableId, tableId), lt(userTableRows.orderKey, anchorKey)))
-      .orderBy(desc(userTableRows.orderKey))
-      .limit(1)
-    prevKey = prev?.orderKey ?? null
-  }
-  return {
-    orderKey: keyBetween(prevKey, anchorKey),
-    position: anchor.position,
-  }
-}
-
-/**
- * Computes fractional `order_key`s for a batch insert. With no `positions`,
- * appends a contiguous run after the current max key. With explicit `positions`
- * (undo restore), keys each row between its pre-shift position neighbors —
- * correct because requested positions are distinct. Caller holds the lock.
- *
- * The explicit-`positions` path is meaningful only when `position` is
- * authoritative (flag off): with the flag on, a saved `position` is a gappy
- * column value, not a visual rank, so feeding it to {@link resolveInsertOrderKey}
- * (which reads `position` as an `OFFSET` rank under the flag) would mint keys at
- * the wrong ranks. Callers needing exact placement under the flag pass
- * `orderKeys` (handled before this function); here we just append a run.
- */
-async function resolveBatchInsertOrderKeys(
-  trx: DbTransaction,
-  tableId: string,
-  count: number,
-  positions?: number[]
-): Promise<string[]> {
-  if (!positions || positions.length === 0 || isTablesFractionalOrderingEnabled) {
-    return nKeysBetween(await maxOrderKey(trx, tableId), null, count)
-  }
-  const keys: string[] = []
-  for (const pos of positions) {
-    keys.push(await resolveInsertOrderKey(trx, tableId, pos))
-  }
-  return keys
-}
-
-/**
- * Inserts a single row in its own transaction. Always assigns a fractional
- * `order_key`. When the fractional-ordering flag is on, `order_key` is
- * authoritative and `position` is a best-effort append (no O(N) shift); when
- * off, `position` is reserved as before (shifting to open the slot). Validation
- * and side-effect dispatch stay with the caller; capacity is enforced by the
- * `increment_user_table_row_count` trigger.
- */
-async function insertOrderedRow(params: {
-  tableId: string
-  workspaceId: string
-  data: RowData
-  rowId: string
-  position?: number
-  afterRowId?: string
-  beforeRowId?: string
-  createdBy?: string
-  now: Date
-}): Promise<{
-  id: string
-  data: RowData
-  position: number
-  orderKey: string | null
-  createdAt: Date
-  updatedAt: Date
-}> {
-  const { tableId, workspaceId, data, rowId, position, afterRowId, beforeRowId, createdBy, now } =
-    params
-  const [row] = await db.transaction(async (trx) => {
-    await setTableTxTimeouts(trx)
-    await acquireRowOrderLock(trx, tableId)
-
-    // Resolve the order key (and a legacy slot position for the flag-off shift
-    // path) from neighbor ids when given, else from the requested position.
-    let orderKey: string
-    let slotPosition = position
-    if (afterRowId || beforeRowId) {
-      const resolved = await resolveInsertByNeighbor(trx, tableId, afterRowId, beforeRowId)
-      orderKey = resolved.orderKey
-      slotPosition = resolved.position
-    } else {
-      orderKey = await resolveInsertOrderKey(trx, tableId, position)
-    }
-
-    let targetPosition: number
-    if (isTablesFractionalOrderingEnabled) {
-      // order_key is authoritative — keep a best-effort, no-shift position.
-      targetPosition = await nextRowPosition(trx, tableId)
-    } else if (slotPosition !== undefined) {
-      const [existing] = await trx
-        .select({ id: userTableRows.id })
-        .from(userTableRows)
-        .where(and(eq(userTableRows.tableId, tableId), eq(userTableRows.position, slotPosition)))
-        .limit(1)
-      if (existing) await shiftRowsUpFrom(trx, tableId, slotPosition)
-      targetPosition = slotPosition
-    } else {
-      targetPosition = await nextRowPosition(trx, tableId)
-    }
-
-    return trx
-      .insert(userTableRows)
-      .values({
-        id: rowId,
-        tableId,
-        workspaceId,
-        data,
-        position: targetPosition,
-        orderKey,
-        createdAt: now,
-        updatedAt: now,
-        ...(createdBy ? { createdBy } : {}),
-      })
-      .returning()
-  })
-  return {
-    id: row.id,
-    data: row.data as RowData,
-    position: row.position,
-    orderKey: row.orderKey,
-    createdAt: row.createdAt,
-    updatedAt: row.updatedAt,
-  }
-}
-
-/**
- * Deletes a single row by id in its own transaction, then closes the positional
- * gap. Returns `false` when no row matched.
- */
-async function deleteOrderedRow(params: {
-  tableId: string
-  rowId: string
-  workspaceId: string
-}): Promise<boolean> {
-  const { tableId, rowId, workspaceId } = params
-  return db.transaction(async (trx) => {
-    await setTableTxTimeouts(trx)
-    const [deleted] = await trx
-      .delete(userTableRows)
-      .where(
-        and(
-          eq(userTableRows.id, rowId),
-          eq(userTableRows.tableId, tableId),
-          eq(userTableRows.workspaceId, workspaceId)
-        )
-      )
-      .returning({ position: userTableRows.position })
-    if (!deleted) return false
-    // Fractional ordering: deleting a row never changes another row's order_key,
-    // so the O(N) position reshift is skipped entirely.
-    if (!isTablesFractionalOrderingEnabled) {
-      await shiftRowsDownAfter(trx, tableId, deleted.position)
-    }
-    return true
-  })
-}
-
-/**
- * Deletes the given row ids in batches within one transaction, then recompacts
- * positions from the earliest deleted slot. Returns the deleted rows (id + prior
- * position). The caller resolves which ids to delete (used by both delete-by-ids
- * and delete-by-filter).
- */
-async function deleteOrderedRowsByIds(params: {
-  tableId: string
-  workspaceId: string
-  rowIds: string[]
-}): Promise<{ id: string; position: number }[]> {
-  const { tableId, workspaceId, rowIds } = params
-  if (rowIds.length === 0) return []
-  return db.transaction(async (trx) => {
-    await setTableTxTimeouts(trx, { statementMs: 60_000 })
-    const deleted: { id: string; position: number }[] = []
-    for (let i = 0; i < rowIds.length; i += TABLE_LIMITS.DELETE_BATCH_SIZE) {
-      const batch = rowIds.slice(i, i + TABLE_LIMITS.DELETE_BATCH_SIZE)
-      const rows = await trx
-        .delete(userTableRows)
-        .where(
-          and(
-            eq(userTableRows.tableId, tableId),
-            eq(userTableRows.workspaceId, workspaceId),
-            inArray(userTableRows.id, batch)
-          )
-        )
-        .returning({ id: userTableRows.id, position: userTableRows.position })
-      deleted.push(...rows)
-    }
-    // Fractional ordering: deletes leave order_key untouched, so no recompaction.
-    if (!isTablesFractionalOrderingEnabled && deleted.length > 0) {
-      const minDeletedPos = deleted.reduce(
-        (min, r) => (r.position < min ? r.position : min),
-        deleted[0].position
-      )
-      await compactPositions(trx, tableId, minDeletedPos)
-    }
-    return deleted
-  })
-}
-
-/**
- * Selects one page of row ids to delete for the async delete-job worker: base scope plus a
- * `created_at <= cutoff` floor (so rows inserted after the job started are never selected) and
- * the caller's optional filter clause. Keyset paginated on `id` via `afterId` so excluded rows
- * (which are skipped, not deleted) still advance the cursor — no OFFSET, no risk of looping on a
- * fully-excluded page.
- */
-export async function selectRowIdPage(params: {
-  tableId: string
-  workspaceId: string
-  cutoff: Date
-  filterClause?: SQL
-  afterId?: string
-  limit: number
-}): Promise<string[]> {
-  const { tableId, workspaceId, cutoff, filterClause, afterId, limit } = params
-  const selectPage = (executor: DbExecutor) =>
-    executor
-      .select({ id: userTableRows.id })
-      .from(userTableRows)
-      .where(
-        and(
-          eq(userTableRows.tableId, tableId),
-          eq(userTableRows.workspaceId, workspaceId),
-          lte(userTableRows.createdAt, cutoff),
-          afterId ? gt(userTableRows.id, afterId) : undefined,
-          filterClause
-        )
-      )
-      .orderBy(asc(userTableRows.id))
-      .limit(limit)
-  // A jsonb filter is unestimatable, so the planner would seq-scan the whole shared relation
-  // per page (12.6s measured) — keep it on the tenant's (table_id, id) index.
-  const rows = filterClause
-    ? await withSeqscanOff(async (trx) => selectPage(trx))
-    : await selectPage(db)
-  return rows.map((r) => r.id)
-}
-
-/**
- * Deletes one page of rows for the async delete-job worker, committing each `DELETE_BATCH_SIZE`
- * chunk in its own short transaction. One statement per transaction bounds how long the
- * statement-level row_count trigger's lock on the definition row is held (a page-wide transaction
- * held it for the entire page, starving concurrent inserts and overrunning `statement_timeout`),
- * and a mid-page failure loses at most one uncommitted batch — the keyset walker (or a task
- * retry) re-walks whatever remains. Skips legacy position compaction: under fractional ordering
- * it's unnecessary, and in the legacy path `position` gaps are harmless — rows still order by
- * position. Returns the count deleted.
- */
-export async function deletePageByIds(
-  tableId: string,
-  workspaceId: string,
-  rowIds: string[]
-): Promise<number> {
-  let deleted = 0
-  for (let i = 0; i < rowIds.length; i += TABLE_LIMITS.DELETE_BATCH_SIZE) {
-    const batch = rowIds.slice(i, i + TABLE_LIMITS.DELETE_BATCH_SIZE)
-    const rows = await db.transaction(async (trx) => {
-      await setTableTxTimeouts(trx, { statementMs: 60_000 })
-      return trx
-        .delete(userTableRows)
-        .where(
-          and(
-            eq(userTableRows.tableId, tableId),
-            eq(userTableRows.workspaceId, workspaceId),
-            inArray(userTableRows.id, batch)
-          )
-        )
-        .returning({ id: userTableRows.id })
-    })
-    deleted += rows.length
-  }
-  return deleted
 }
 
 /**
@@ -3071,248 +2425,12 @@ export async function getRowById(
   }
 }
 
-/**
- * Derive automatic clears + cancellation candidates from a row's data patch.
- *
- * Walks `schema.workflowGroups` left-to-right with a propagating `dirtied`
- * column set. For each group whose deps overlap the dirty set, decide to
- * clear (terminal exec) or cancel+rerun (in-flight exec), then add the
- * group's outputs to the dirty set so later groups in the chain see them
- * as dirty too. This models transitive dep chains as a single forward pass —
- * editing column A propagates through group 1 (deps on A) to group 2 (deps
- * on group 1's output) without explicit DAG traversal.
- *
- * Returns:
- * - `executionsPatch`: caller's patch + nulls for cleared groups (or
- *   undefined if nothing applied).
- * - `inFlightDownstreamGroups`: groups whose dep was dirtied and that are
- *   currently in-flight. Cancel-and-restart is the caller's job.
- *
- * Assumption: `workflowGroups[]` is in topological order — a group's deps
- * may only reference columns to its left (enforced by `workflow-sidebar`'s
- * "Run after" picker + the reorder scrub via `stripGroupDeps`). Violating
- * this would silently miss the propagation.
- */
-function deriveExecClearsForDataPatch(
-  dataPatch: RowData,
-  schema: TableSchema,
-  existingExecutions: RowExecutions,
-  callerPatch: Record<string, RowExecutionMetadata | null> | undefined,
-  mergedData: RowData
-): {
-  executionsPatch: Record<string, RowExecutionMetadata | null> | undefined
-  inFlightDownstreamGroups: string[]
-} {
-  const dirtied = new Set(Object.keys(dataPatch))
-  const groupsToClear = new Set<string>()
-  const inFlightDownstreamGroups: string[] = []
-
-  // Own-output clears: when the user wipes a workflow output column, drop
-  // that group's exec entry so the auto-fire reactor re-arms the cell.
-  // Also flags the cleared output column as dirty so transitive downstream
-  // groups see it.
-  for (const [columnId, value] of Object.entries(dataPatch)) {
-    const cleared = value === null || value === undefined || value === ''
-    if (!cleared) continue
-    const col = schema.columns.find((c) => getColumnId(c) === columnId)
-    if (col?.workflowGroupId) groupsToClear.add(col.workflowGroupId)
-  }
-
-  // Left-to-right walk, propagating dirty columns forward.
-  const groups = schema.workflowGroups ?? []
-  const afterRow = { data: mergedData } as TableRow
-  for (const group of groups) {
-    const deps = group.dependencies?.columns ?? []
-    const depMatched = deps.some((d) => dirtied.has(d))
-    if (!depMatched) continue
-
-    // A dep column changed, but if the group's deps are no longer satisfied
-    // after the patch — a checkbox was unchecked or a text dep cleared — there's
-    // nothing to recompute. Leave the prior result alone instead of re-arming or
-    // cancelling it; only checking a box / filling a dep drives downstream work.
-    if (!areGroupDepsSatisfied(group, afterRow)) continue
-
-    const exec = existingExecutions[group.id]
-    if (exec) {
-      const status = exec.status
-      if (status === 'completed' || status === 'error' || status === 'cancelled') {
-        groupsToClear.add(group.id)
-      } else if (status === 'queued' || status === 'running' || status === 'pending') {
-        inFlightDownstreamGroups.push(group.id)
-      }
-    } else {
-      // No exec entry yet — `mode: 'new'` already covers this group. We
-      // still propagate the dirty signal forward so later groups in the
-      // chain see this group's outputs as dirty too.
-      groupsToClear.add(group.id)
-    }
-
-    // Propagate: this group is about to be re-computed, so groups whose
-    // deps reference its output columns are also dirty.
-    for (const out of group.outputs) dirtied.add(out.columnName)
-  }
-
-  if (groupsToClear.size === 0) {
-    return { executionsPatch: callerPatch, inFlightDownstreamGroups }
-  }
-  const merged: Record<string, RowExecutionMetadata | null> = { ...(callerPatch ?? {}) }
-  for (const gid of groupsToClear) {
-    if (!(gid in merged)) merged[gid] = null
-  }
-  return { executionsPatch: merged, inFlightDownstreamGroups }
-}
-
 /** Internal: thrown inside `db.transaction` to roll back when the executions
  *  guard rejects a write. The outer `.catch` translates it into a `null` return. */
 class GuardRejected extends Error {
   constructor() {
     super('cell-write guard rejected')
   }
-}
-
-/** Merges an `executionsPatch` into the row's existing executions blob. */
-function applyExecutionsPatch(
-  existing: RowExecutions,
-  patch: Record<string, RowExecutionMetadata | null> | undefined
-): RowExecutions {
-  if (!patch) return existing
-  const next: RowExecutions = { ...existing }
-  for (const [gid, value] of Object.entries(patch)) {
-    if (value === null) {
-      delete next[gid]
-    } else {
-      next[gid] = value
-    }
-  }
-  return next
-}
-
-/**
- * Writes a per-group execution patch for one row against the `tableRowExecutions`
- * sidecar. Non-null values upsert into the table; nulls delete the entry. When
- * `guard` is set, the upsert is gated to:
- *  - reject if a `cancelled` row for the same execution already exists, and
- *  - reject if the row exists but is owned by a different executionId
- *    (with carve-outs for missing rows and null executionIds — the dispatcher's
- *    pre-batch `pending` stamp leaves executionId unset so the first cell-task
- *    can claim).
- *
- * Returns `'guard-rejected'` when the guarded group's upsert affected 0 rows
- * (callers signal failure to the cell-task path). Returns `'wrote'` otherwise.
- */
-async function writeExecutionsPatch(
-  trx: DbOrTx,
-  tableId: string,
-  rowId: string,
-  patch: Record<string, RowExecutionMetadata | null> | undefined,
-  guard?: { groupId: string; executionId: string }
-): Promise<'wrote' | 'guard-rejected'> {
-  if (!patch) return 'wrote'
-  const entries = Object.entries(patch)
-  if (entries.length === 0) return 'wrote'
-
-  for (const [gid, value] of entries) {
-    if (value === null) {
-      await trx
-        .delete(tableRowExecutions)
-        .where(and(eq(tableRowExecutions.rowId, rowId), eq(tableRowExecutions.groupId, gid)) as SQL)
-      continue
-    }
-    const insertValues = {
-      tableId,
-      rowId,
-      groupId: gid,
-      status: value.status,
-      executionId: value.executionId,
-      jobId: value.jobId,
-      workflowId: value.workflowId,
-      error: value.error,
-      runningBlockIds: value.runningBlockIds ?? [],
-      blockErrors: value.blockErrors ?? {},
-      cancelledAt: value.cancelledAt ? new Date(value.cancelledAt) : null,
-      updatedAt: new Date(),
-    } as const
-
-    const isGuarded = guard && guard.groupId === gid
-    if (isGuarded) {
-      // Gate by guard semantics. The original JSONB guard had two AND'd
-      // clauses; we collapse them onto the upsert's WHERE so a non-matching
-      // existing row leaves the table untouched and we observe 0 affected.
-      const guardExecutionId = guard.executionId
-      const updated = await trx
-        .insert(tableRowExecutions)
-        .values(insertValues)
-        .onConflictDoUpdate({
-          target: [tableRowExecutions.rowId, tableRowExecutions.groupId],
-          set: {
-            status: insertValues.status,
-            executionId: insertValues.executionId,
-            jobId: insertValues.jobId,
-            workflowId: insertValues.workflowId,
-            error: insertValues.error,
-            runningBlockIds: insertValues.runningBlockIds,
-            blockErrors: insertValues.blockErrors,
-            cancelledAt: insertValues.cancelledAt,
-            updatedAt: insertValues.updatedAt,
-          },
-          where: and(
-            // Reject any guarded worker write when the cell is `cancelled` — a
-            // stop click wrote it authoritatively. SQL mirror of `isExecCancelled`
-            // (deps.ts). Status-only (not executionId-scoped): the cancel can
-            // only carry the pre-stamp's executionId (often null), so matching on
-            // id would let the worker's real-id claim resurrect a killed cell.
-            sql`${tableRowExecutions.status} <> 'cancelled'`,
-            // Stale-worker: the cell's active run has moved on. Carve-outs
-            // permit a fresh worker to take over when the row's executionId
-            // is unset (dispatcher's pre-batch `pending` stamp).
-            sql`(${tableRowExecutions.executionId} IS NULL OR ${tableRowExecutions.executionId} = ${guardExecutionId})`
-          ) as SQL,
-        })
-        .returning({ rowId: tableRowExecutions.rowId })
-      if (updated.length === 0) return 'guard-rejected'
-      continue
-    }
-
-    await trx
-      .insert(tableRowExecutions)
-      .values(insertValues)
-      .onConflictDoUpdate({
-        target: [tableRowExecutions.rowId, tableRowExecutions.groupId],
-        set: {
-          status: insertValues.status,
-          executionId: insertValues.executionId,
-          jobId: insertValues.jobId,
-          workflowId: insertValues.workflowId,
-          error: insertValues.error,
-          runningBlockIds: insertValues.runningBlockIds,
-          blockErrors: insertValues.blockErrors,
-          cancelledAt: insertValues.cancelledAt,
-          updatedAt: insertValues.updatedAt,
-        },
-      })
-  }
-
-  return 'wrote'
-}
-
-/**
- * Strips the given workflow group ids from every row's executions on a table —
- * used by the column / group delete paths so stale running/queued exec records
- * don't linger and inflate counters after the group is gone. The caller wraps
- * in their own transaction.
- */
-async function stripGroupExecutions(
-  trx: DbOrTx,
-  tableId: string,
-  groupIds: Iterable<string>
-): Promise<void> {
-  const ids = Array.from(new Set(groupIds))
-  if (ids.length === 0) return
-  await trx
-    .delete(tableRowExecutions)
-    .where(
-      and(eq(tableRowExecutions.tableId, tableId), inArray(tableRowExecutions.groupId, ids)) as SQL
-    )
 }
 
 /**
